+++ tests/test_cookbook.py
@@ -1,29 +1,41 @@
 from src.cookbook import cookbook
 from click.testing import CliRunner
+from os import system
 
 
-def test_download_recipe():
+def test_save():
     runner = CliRunner()
     url = "http://example.com/"
-    filename = 'Example'
+    filename = "Example"
 
-    result = runner.invoke(cookbook.download_recipe, ['-s', url, '-n', filename])
-    
+    result = runner.invoke(cookbook.save, ["-u", url, "-n", filename])
+
     assert result.exit_code == 0
-    assert result.output == 'Example saved\n'
+    assert result.output == "Example saved\n"
 
     filepath = f"/home/vedran/Documents/{filename}.html"
 
     with open(filepath) as f:
         content = f.read()
 
+    system(f"rm -f {filepath}")
+
     assert len(content) > 0
 
 
-    def test_list_recipes():
-        runner = CliRunner()
-        result = runner.invoke(cookbook.list_recipes)
+def test_browse():
+    runner = CliRunner()
+    result = runner.invoke(cookbook.browse)
 
-        #captured_out = result.output.split('\n')
+    # captured_out = result.output.split('\n')
 
-        assert result.exit_code == 0
+    assert result.exit_code == 0
+
+
+def test_view():
+    filepath = f"/home/vedran/Documents/Example.html"
+
+    runner = CliRunner()
+    result = runner.invoke(cookbook.view, [filepath])
+    system(f"rm -f {filepath}")
+    assert result.exit_code == 0